refactor: add GameBoard and GamePiece contexts with custom hooks for managing game elements
